+++ src/sql/util.py
@@ -1,0 +1,31 @@
+def convert_to_scientific(value):
+    """
+    Converts value to scientific notation if necessary
+
+    Parameters
+    ----------
+    value : any
+        Value to format.
+    """
+    if (
+        isinstance(value, (int, float))
+        and not isinstance(value, bool)
+        and _is_long_number(value)
+    ):
+        new_value = "{:,.3e}".format(value)
+
+    else:
+        new_value = value
+
+    return new_value
+
+
+def _is_long_number(num) -> bool:
+    """
+    Checks if num's digits > 10
+    """
+    if "." in str(num):
+        split_by_decimal = str(num).split(".")
+        if len(split_by_decimal[0]) > 10 or len(split_by_decimal[1]) > 10:
+            return True
+    return False